+++ tests/glob.spec.ts
@@ -6,6 +6,8 @@ const controllerConstructor = mod.default
 const withoutDefaultExport = { }
 
 const modules = {
+  '../components/ui/home_component/home_component_controller.js': mod,
+  '../components/button_component/button_component_controller.js': mod,
   '../components/home_controller.js': mod,
   '../components/page/controller.js': mod,
   '../controllers/hello_controller.js': mod,
@@ -18,6 +20,8 @@ const modules = {
 
 test('definitionsFromGlob', () => {
   expect(definitionsFromGlob(modules)).toEqual([
+    { identifier: 'ui--home-component--home-component', controllerConstructor },
+    { identifier: 'button-component--button-component', controllerConstructor },
     { identifier: 'home', controllerConstructor },
     { identifier: 'page', controllerConstructor },
     { identifier: 'hello', controllerConstructor },
@@ -27,3 +31,18 @@ test('definitionsFromGlob', () => {
     { identifier: 'image--blur', controllerConstructor },
   ])
 })
+
+test('definitionsFromGlob with nestedMode', () => {
+  expect(definitionsFromGlob(modules, true)).toEqual([
+    { identifier: 'ui--home-component', controllerConstructor },
+    { identifier: 'button-component', controllerConstructor },
+    { identifier: 'home', controllerConstructor },
+    { identifier: 'page', controllerConstructor },
+    { identifier: 'hello', controllerConstructor },
+    { identifier: 'image--reveal', controllerConstructor },
+    { identifier: 'dashboard', controllerConstructor },
+    { identifier: 'image--reveal', controllerConstructor },
+    { identifier: 'image--blur', controllerConstructor },
+  ])
+})
+